Move API classes again
