--- src/api/java/ladysnake/requiem/api/v1/player/package-info.java
@@ -1,9 +1,0 @@
-@FieldsAreNonnullByDefault
-@MethodsReturnNonnullByDefault
-@ParametersAreNonnullByDefault
-package ladysnake.requiem.api.v1.player;
-
-import ladysnake.requiem.api.v1.annotation.FieldsAreNonnullByDefault;
-import ladysnake.requiem.api.v1.annotation.MethodsReturnNonnullByDefault;
-
-import javax.annotation.ParametersAreNonnullByDefault;
