+++ src/main/java/ladysnake/requiem/common/impl/movement/PlayerMovementAlterer.java
@@ -19,7 +19,7 @@ package ladysnake.requiem.common.impl.movement;
 
 import ladysnake.requiem.api.v1.entity.MovementAlterer;
 import ladysnake.requiem.api.v1.entity.MovementConfig;
-import ladysnake.requiem.api.v1.player.RequiemPlayer;
+import ladysnake.requiem.api.v1.RequiemPlayer;
 import net.minecraft.client.network.packet.PlayerAbilitiesS2CPacket;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.LivingEntity;
