+++ src/main/java/ladysnake/requiem/mixin/client/entity/ClientEntityMixin.java
@@ -17,7 +17,7 @@
  */
 package ladysnake.requiem.mixin.client.entity;
 
-import ladysnake.requiem.api.v1.player.RequiemPlayer;
+import ladysnake.requiem.api.v1.RequiemPlayer;
 import net.minecraft.entity.Entity;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
