+++ src/main/java/ladysnake/requiem/mixin/possession/entity/mob/ZombieEntityMixin.java
@@ -17,7 +17,7 @@
  */
 package ladysnake.requiem.mixin.possession.entity.mob;
 
-import ladysnake.requiem.api.v1.player.RequiemPlayer;
+import ladysnake.requiem.api.v1.RequiemPlayer;
 import ladysnake.requiem.api.v1.possession.Possessable;
 import ladysnake.requiem.api.v1.possession.PossessionComponent;
 import net.minecraft.entity.EntityType;
