+++ src/main/java/ladysnake/requiem/mixin/possession/entity/passive/HorseBaseEntityMixin.java
@@ -17,7 +17,7 @@
  */
 package ladysnake.requiem.mixin.possession.entity.passive;
 
-import ladysnake.requiem.api.v1.player.RequiemPlayer;
+import ladysnake.requiem.api.v1.RequiemPlayer;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.LivingEntity;
